# This submission is used to improve the performance of the inpaint algorithm for 3 channels images(RGB or BGR)