@easonycwang 2021-09-24
> Thank you for contribution!
> 
> This patch should go into [3.4 branch](https://github.com/opencv/opencv/wiki/Branches) first.
> We will merge changes from 3.4 into master regularly (weekly/bi-weekly).
> 
> Please:
> 
> * change "base" branch of this PR: master => 3.4 (use "Edit" button near PR title)
> * rebase your commits from master onto 3.4 branch. For example:
>   `git rebase -i --onto upstream/3.4 upstream/master`
>   (check list of your commits, save and quit (Esc + "wq" + Enter)
>   where `upstream` is configured by following [this GitHub guide](https://help.github.com/articles/configuring-a-remote-for-a-fork/) and fetched (`git fetch upstream`).
> * push rebased commits into source branch of your fork (with `--force` option)
> 
> Note: no needs to re-open PR, apply changes "inplace".